Added Gravatar/Robohash profile images
Profile images use Gravatar images if they are configured for the users
email, and if not a default Robohash image will be used.

Added avatar hash to db to help performance. Pre-computing the hash used
for avatars and storing it in the database to avoid computing the hash
for every avatar displayed on a page.